move server files
