+++ server/CircuitParts/Ports.ts
@@ -1,10 +1,10 @@
-import type { ProblemCategory } from '#/circuit/BothScript/CircuitInfo'
+import type { ProblemCategory } from '../../src/circuit/BothScript/CircuitInfo'
 import { CircuitParts } from '../CircuitParts'
-export class CPU extends CircuitParts {
-  public sizeX = 2
+export class Ports extends CircuitParts {
+  public sizeX = 4
   public sizeY = 2
   public wirePointsInts: number[][] = new Array(this.sizeY * this.sizeX)
-  public problemCategory: ProblemCategory = 'None'
+  public problemCategory: ProblemCategory = 'OSINT'
   constructor(isBig: boolean) {
     super(isBig)
     this.wirePointsInts.forEach(
@@ -16,6 +16,14 @@ export class CPU extends CircuitParts {
       1, 1, 1, 1, 1, 1, 1
     ]
     this.wirePointsInts[1 + 0 * this.sizeX] = [
+      0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0,
+      0, 0, 0, 0, 0, 0, 0
+    ]
+    this.wirePointsInts[2 + 0 * this.sizeX] = [
+      0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0,
+      0, 0, 0, 0, 0, 0, 0
+    ]
+    this.wirePointsInts[3 + 0 * this.sizeX] = [
       0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0,
       0, 0, 0, 0, 0, 0, 0
     ]
@@ -24,6 +32,14 @@ export class CPU extends CircuitParts {
       1, 1, 1, 1, 1, 1, 1
     ]
     this.wirePointsInts[1 + 1 * this.sizeX] = [
+      1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+      0, 0, 0, 0, 0, 0, 0
+    ]
+    this.wirePointsInts[2 + 1 * this.sizeX] = [
+      1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+      0, 0, 0, 0, 0, 0, 0
+    ]
+    this.wirePointsInts[3 + 1 * this.sizeX] = [
       1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0,
       0, 0, 0, 0, 0, 0, 0
     ]
