+++ server/CircuitParts/SSD.ts
@@ -1,10 +1,10 @@
-import type { ProblemCategory } from '#/circuit/BothScript/CircuitInfo'
+import type { ProblemCategory } from '../../src/circuit/BothScript/CircuitInfo'
 import { CircuitParts } from '../CircuitParts'
-export class Ports extends CircuitParts {
-  public sizeX = 4
+export class SSD extends CircuitParts {
+  public sizeX = 8
   public sizeY = 2
   public wirePointsInts: number[][] = new Array(this.sizeY * this.sizeX)
-  public problemCategory: ProblemCategory = 'OSINT'
+  public problemCategory: ProblemCategory = 'Misc'
   constructor(isBig: boolean) {
     super(isBig)
     this.wirePointsInts.forEach(
@@ -24,6 +24,22 @@ export class Ports extends CircuitParts {
       0, 0, 0, 0, 0, 0, 0
     ]
     this.wirePointsInts[3 + 0 * this.sizeX] = [
+      0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0,
+      0, 0, 0, 0, 0, 0, 0
+    ]
+    this.wirePointsInts[4 + 0 * this.sizeX] = [
+      0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0,
+      0, 0, 0, 0, 0, 0, 0
+    ]
+    this.wirePointsInts[5 + 0 * this.sizeX] = [
+      0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0,
+      0, 0, 0, 0, 0, 0, 0
+    ]
+    this.wirePointsInts[6 + 0 * this.sizeX] = [
+      0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0,
+      0, 0, 0, 0, 0, 0, 0
+    ]
+    this.wirePointsInts[7 + 0 * this.sizeX] = [
       0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0,
       0, 0, 0, 0, 0, 0, 0
     ]
@@ -40,6 +56,22 @@ export class Ports extends CircuitParts {
       0, 0, 0, 0, 0, 0, 0
     ]
     this.wirePointsInts[3 + 1 * this.sizeX] = [
+      1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+      0, 0, 0, 0, 0, 0, 0
+    ]
+    this.wirePointsInts[4 + 1 * this.sizeX] = [
+      1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+      0, 0, 0, 0, 0, 0, 0
+    ]
+    this.wirePointsInts[5 + 1 * this.sizeX] = [
+      1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+      0, 0, 0, 0, 0, 0, 0
+    ]
+    this.wirePointsInts[6 + 1 * this.sizeX] = [
+      1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+      0, 0, 0, 0, 0, 0, 0
+    ]
+    this.wirePointsInts[7 + 1 * this.sizeX] = [
       1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0,
       0, 0, 0, 0, 0, 0, 0
     ]
