+++ server/CircuitParts/TipSet.ts
@@ -1,0 +1,31 @@
+import type { ProblemCategory } from '../../src/circuit/BothScript/CircuitInfo'
+import { CircuitParts } from '../CircuitParts'
+export class TipSet extends CircuitParts {
+  public sizeX = 2
+  public sizeY = 2
+  public wirePointsInts: number[][] = new Array(this.sizeY * this.sizeX)
+  public problemCategory: ProblemCategory = 'Shell'
+  constructor(isBig: boolean) {
+    super(isBig)
+    this.wirePointsInts.forEach(
+      (_, i) => (this.wirePointsInts[i] = new Array(32))
+    )
+    //x=0,y=0のcellのwirePoints
+    this.wirePointsInts[0 + 0 * this.sizeX] = [
+      0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1,
+      1, 1, 1, 1, 1, 1, 1
+    ]
+    this.wirePointsInts[1 + 0 * this.sizeX] = [
+      0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0,
+      0, 0, 0, 0, 0, 0, 0
+    ]
+    this.wirePointsInts[0 + 1 * this.sizeX] = [
+      1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1,
+      1, 1, 1, 1, 1, 1, 1
+    ]
+    this.wirePointsInts[1 + 1 * this.sizeX] = [
+      1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+      0, 0, 0, 0, 0, 0, 0
+    ]
+  }
+}
