--- src/circuit/ServerScript/CircuitParts/Memory.ts
@@ -1,79 +1,0 @@
-import type { ProblemCategory } from '#/circuit/BothScript/CircuitInfo'
-import { CircuitParts } from '../CircuitParts'
-export class Memory extends CircuitParts {
-  public sizeX = 8
-  public sizeY = 2
-  public wirePointsInts: number[][] = new Array(this.sizeY * this.sizeX)
-  public problemCategory: ProblemCategory = 'PPC'
-  constructor(isBig: boolean) {
-    super(isBig)
-    this.wirePointsInts.forEach(
-      (_, i) => (this.wirePointsInts[i] = new Array(32))
-    )
-    //x=0,y=0のcellのwirePoints
-    this.wirePointsInts[0 + 0 * this.sizeX] = [
-      0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1,
-      1, 1, 1, 1, 1, 1, 1
-    ]
-    this.wirePointsInts[1 + 0 * this.sizeX] = [
-      0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0,
-      0, 0, 0, 0, 0, 0, 0
-    ]
-    this.wirePointsInts[2 + 0 * this.sizeX] = [
-      0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0,
-      0, 0, 0, 0, 0, 0, 0
-    ]
-    this.wirePointsInts[3 + 0 * this.sizeX] = [
-      0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0,
-      0, 0, 0, 0, 0, 0, 0
-    ]
-    this.wirePointsInts[4 + 0 * this.sizeX] = [
-      0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0,
-      0, 0, 0, 0, 0, 0, 0
-    ]
-    this.wirePointsInts[5 + 0 * this.sizeX] = [
-      0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0,
-      0, 0, 0, 0, 0, 0, 0
-    ]
-    this.wirePointsInts[6 + 0 * this.sizeX] = [
-      0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0,
-      0, 0, 0, 0, 0, 0, 0
-    ]
-    this.wirePointsInts[7 + 0 * this.sizeX] = [
-      0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0,
-      0, 0, 0, 0, 0, 0, 0
-    ]
-    this.wirePointsInts[0 + 1 * this.sizeX] = [
-      1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1,
-      1, 1, 1, 1, 1, 1, 1
-    ]
-    this.wirePointsInts[1 + 1 * this.sizeX] = [
-      1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-      0, 0, 0, 0, 0, 0, 0
-    ]
-    this.wirePointsInts[2 + 1 * this.sizeX] = [
-      1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-      0, 0, 0, 0, 0, 0, 0
-    ]
-    this.wirePointsInts[3 + 1 * this.sizeX] = [
-      1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-      0, 0, 0, 0, 0, 0, 0
-    ]
-    this.wirePointsInts[4 + 1 * this.sizeX] = [
-      1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-      0, 0, 0, 0, 0, 0, 0
-    ]
-    this.wirePointsInts[5 + 1 * this.sizeX] = [
-      1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-      0, 0, 0, 0, 0, 0, 0
-    ]
-    this.wirePointsInts[6 + 1 * this.sizeX] = [
-      1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-      0, 0, 0, 0, 0, 0, 0
-    ]
-    this.wirePointsInts[7 + 1 * this.sizeX] = [
-      1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-      0, 0, 0, 0, 0, 0, 0
-    ]
-  }
-}
